@@ -6,12 +6,11 @@ import net.ninjacat.semblance.data.collections.LispValue;
 import net.ninjacat.semblance.data.collections.NilCollection;
 import net.ninjacat.semblance.data.collections.SList;
 import net.ninjacat.semblance.evaluator.Context;
-import net.ninjacat.smooth.functions.Func;
-import net.ninjacat.smooth.iterators.Iter;
 
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
+import java.util.stream.Collectors;
 
 import static net.ninjacat.semblance.utils.Values.asCollection;
 
@@ -45,12 +44,9 @@ public class Zip extends BuiltInFunction {
         final LispCollection main = asCollection(evaluated.head());
         final LispCollection rest = evaluated.tail();
 
-        final List<Iterator<LispValue>> iters = Iter.of(rest.getCollection()).map(new Func<Iterator<LispValue>, LispValue>() {
-            @Override
-            public Iterator<LispValue> apply(final LispValue value) {
-                return asCollection(value).iterator();
-            }
-        }).toList();
+        final List<Iterator<LispValue>> iters = rest.stream()
+                .map(value -> asCollection(value).iterator())
+                .collect(Collectors.toList());
 
         final List<LispValue> result = new ArrayList<>();
 
