@@ -1,5 +1,6 @@
 package net.ninjacat.semblance.builtin.spforms;
 
+import com.google.common.collect.ImmutableList;
 import net.ninjacat.semblance.data.Constants.HiddenFunctions;
 import net.ninjacat.semblance.data.SymbolAtom;
 import net.ninjacat.semblance.data.callables.SpecialForm;
@@ -8,14 +9,12 @@ import net.ninjacat.semblance.data.collections.LispValue;
 import net.ninjacat.semblance.data.collections.SList;
 import net.ninjacat.semblance.errors.runtime.UnboundSymbolException;
 import net.ninjacat.semblance.evaluator.Context;
-import net.ninjacat.smooth.collections.Lists;
-import net.ninjacat.smooth.iterators.Collectors;
-import net.ninjacat.smooth.iterators.Iter;
-import net.ninjacat.smooth.utils.Option;
 
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
+import java.util.Optional;
+import java.util.stream.Collectors;
 
 import static net.ninjacat.semblance.utils.Values.*;
 
@@ -33,13 +32,13 @@ public class Backquote extends SpecialForm {
         super("backquote", "&rest", "body");
     }
 
-    private static boolean isUnquote(final LispValue func) {
-        return func.equals(HiddenFunctions.COMMA);
-    }
-
     @Override
     public LispValue apply(final Context context, final LispCollection parameters) {
         return expandBQ(parameters, context);
+    }
+
+    private static boolean isUnquote(final LispValue func) {
+        return func.equals(HiddenFunctions.COMMA);
     }
 
     private LispCollection expandBQ(final LispCollection sExpr, final Context parameterContext) {
@@ -102,15 +101,15 @@ public class Backquote extends SpecialForm {
 
     private Collection<LispValue> expandParameter(final LispValue item, final Context parameterContext) {
         final SymbolAtom actualName = asSymbol(item);
-        final Option<LispValue> value = parameterContext.findSymbol(actualName);
+        final Optional<LispValue> value = parameterContext.findSymbol(actualName);
         if (value.isPresent()) {
-            return Lists.of(value.get());
+            return ImmutableList.of(value.get());
         } else {
             throw new UnboundSymbolException(asSymbol(item), getSourceInfo(item));
         }
     }
 
     private Collection<LispValue> expandList(final LispValue value) {
-        return Iter.of(asCollection(value).iterator()).collectWith(Collectors.<LispValue>arrayList());
+        return asCollection(value).stream().collect(Collectors.toList());
     }
 }
