@@ -3,8 +3,6 @@ package net.ninjacat.semblance.builtin.spforms.comparison;
 import net.ninjacat.semblance.data.SymbolAtom;
 import net.ninjacat.semblance.data.collections.LispCollection;
 import net.ninjacat.semblance.data.collections.LispValue;
-import net.ninjacat.smooth.functions.Predicate;
-import net.ninjacat.smooth.iterators.Iter;
 
 /**
  * Equals evaluator
@@ -22,10 +20,8 @@ public class Equal extends BaseEqual {
     public LispValue doApply(final LispCollection parameters) {
         final LispValue first = parameters.head();
 
-        final Predicate<LispValue> comparator = createComparatorFor(first);
-
         return SymbolAtom.fromBoolean(
-                Iter.of(parameters.tail().iterator()).all(comparator)
+                parameters.tail().stream().allMatch(createComparatorFor(first))
         );
     }
 
