@@ -3,9 +3,6 @@ package net.ninjacat.semblance.builtin.spforms.comparison;
 import net.ninjacat.semblance.data.SymbolAtom;
 import net.ninjacat.semblance.data.collections.LispCollection;
 import net.ninjacat.semblance.data.collections.LispValue;
-import net.ninjacat.smooth.functions.Predicate;
-import net.ninjacat.smooth.functions.Predicates;
-import net.ninjacat.smooth.iterators.Iter;
 
 /**
  * != evaluator.
@@ -22,11 +19,8 @@ public class NotEqual extends BaseEqual {
     @Override
     public LispValue doApply(final LispCollection parameters) {
         final LispValue first = parameters.head();
-
-        final Predicate<LispValue> comparator = Predicates.not(createComparatorFor(first));
-
         return SymbolAtom.fromBoolean(
-                Iter.of(parameters.tail().iterator()).all(comparator)
+                parameters.tail().stream().noneMatch(createComparatorFor(first))
         );
     }
 }
