@@ -1,12 +1,13 @@
 package net.ninjacat.semblance.builtin.spforms;
 
+import com.google.common.collect.ImmutableList;
 import net.ninjacat.semblance.data.collections.LispValue;
 import net.ninjacat.semblance.errors.runtime.FileNotFoundException;
 import net.ninjacat.semblance.evaluator.Context;
 import net.ninjacat.semblance.evaluator.RootContext;
-import net.ninjacat.smooth.collections.Lists;
-import net.ninjacat.smooth.utils.Option;
 import org.junit.Test;
+
+import java.util.Optional;
 
 import static net.ninjacat.semblance.utils.Values.*;
 import static org.hamcrest.core.Is.is;
@@ -25,10 +26,10 @@ public class IncludeTest {
     public void testShouldLoadDefinitionsFromFile() throws Exception {
         final Include include = new Include();
         final Context context = new RootContext();
-        context.setSourceFolders(Lists.of("jar:"));
+        context.setSourceFolders(ImmutableList.of("jar:"));
         include.apply(context, list(string("test_program")));
 
-        final Option<LispValue> addOp = context.findSymbol(symbol("add"));
+        final Optional<LispValue> addOp = context.findSymbol(symbol("add"));
         assertThat("Should bind add function in current context", addOp.isPresent(), is(true));
     }
 }
