@@ -2,8 +2,9 @@ package net.ninjacat.semblance.evaluator;
 
 import net.ninjacat.semblance.data.LispCallable;
 import net.ninjacat.semblance.data.collections.LispValue;
-import net.ninjacat.smooth.utils.Option;
 import org.junit.Test;
+
+import java.util.Optional;
 
 import static net.ninjacat.semblance.utils.Values.symbol;
 import static org.hamcrest.Matchers.instanceOf;
@@ -16,7 +17,7 @@ public class RootContextTest {
     public void testShouldLoadStandardLibrary() throws Exception {
         final RootContext context = new RootContext();
 
-        final Option<LispValue> defun = context.findSymbol(symbol("defun"));
+        final Optional<LispValue> defun = context.findSymbol(symbol("defun"));
 
         assertThat(defun.isPresent(), is(true));
         assertThat(defun.get(), instanceOf(LispCallable.class));
