@@ -7,9 +7,9 @@ import net.ninjacat.semblance.data.collections.LispValue;
 import net.ninjacat.semblance.debug.SourceInfo;
 import net.ninjacat.semblance.evaluator.Context;
 import net.ninjacat.semblance.java.types.CallHelpers;
-import net.ninjacat.smooth.utils.Option;
 
 import java.lang.reflect.Field;
+import java.util.Optional;
 
 import static java.lang.String.format;
 import static net.ninjacat.semblance.utils.Values.*;
@@ -45,7 +45,7 @@ public class JavaGetField extends SpecialForm {
 
     private static LispValue getStaticField(final SymbolAtom staticField) {
         final StaticReference reference = new StaticReference(staticField);
-        final Option<Field> field = reference.getField();
+        final Optional<Field> field = reference.getField();
         if (!field.isPresent()) {
             throw new JavaInteropException(format("Failed to find static field %s", staticField.repr()));
         }
