@@ -5,9 +5,9 @@ import net.ninjacat.semblance.data.collections.LispCollection;
 import net.ninjacat.semblance.data.collections.LispValue;
 import net.ninjacat.semblance.debug.SourceInfo;
 import net.ninjacat.semblance.evaluator.Context;
-import net.ninjacat.smooth.utils.Option;
 
 import java.lang.reflect.Method;
+import java.util.Optional;
 
 import static net.ninjacat.semblance.java.types.CallHelpers.*;
 import static net.ninjacat.semblance.utils.Values.asSymbol;
@@ -34,7 +34,7 @@ public class JavaStaticCall extends SpecialForm {
         final LispCollection params = context.evaluateList(parameters.tail());
 
         final StaticReference reference = new StaticReference(staticName);
-        final Option<Method> method = findMatchingMethod(reference.getMethods(), params);
+        final Optional<Method> method = findMatchingMethod(reference.getMethods(), params);
 
         if (!method.isPresent()) {
             throw new JavaInteropException(String.format("Cannot find method %s.", staticName));
