@@ -1,25 +1,24 @@
 package net.ninjacat.semblance.java.types;
 
+import com.google.common.collect.ImmutableSet;
 import net.ninjacat.semblance.data.collections.LispCollection;
 import net.ninjacat.semblance.data.collections.LispValue;
 import net.ninjacat.semblance.debug.SourceInfo;
 import net.ninjacat.semblance.java.JavaConvertible;
 import net.ninjacat.semblance.java.JavaInteropException;
+import net.ninjacat.semblance.utils.Maps;
 import net.ninjacat.semblance.utils.Values;
-import net.ninjacat.smooth.collections.Collect;
-import net.ninjacat.smooth.collections.Maps;
-import net.ninjacat.smooth.functions.Func;
-import net.ninjacat.smooth.functions.Predicate;
-import net.ninjacat.smooth.iterators.Iter;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import java.lang.reflect.ParameterizedType;
 import java.lang.reflect.Type;
+import java.util.Arrays;
 import java.util.Collection;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
+import java.util.function.Function;
+import java.util.stream.Collectors;
 
 /**
  * Compatibility handler for collections
@@ -36,14 +35,11 @@ public class ListTypeCompatibility implements TypeCompatibility {
             String[].class, String.class
     );
 
-    private static final Func<Object, LispValue> SIMPLE_VALUE_CONVERTER = new Func<Object, LispValue>() {
-        @Override
-        public Object apply(final LispValue item) {
-            if (item instanceof JavaConvertible) {
-                return ((JavaConvertible) item).asJavaObject();
-            } else {
-                throw new JavaInteropException("Cannot convert " + item + " into Java object", SourceInfo.UNKNOWN);
-            }
+    private static final Function<LispValue, Object> SIMPLE_VALUE_CONVERTER = item -> {
+        if (item instanceof JavaConvertible) {
+            return ((JavaConvertible) item).asJavaObject();
+        } else {
+            throw new JavaInteropException("Cannot convert " + item + " into Java object", SourceInfo.UNKNOWN);
         }
     };
 
@@ -77,7 +73,8 @@ public class ListTypeCompatibility implements TypeCompatibility {
 
     private static boolean isClassCompatible(@Nonnull final Class<?> javaType) {
         return javaType.isArray() ||
-                Iter.of(javaType.getInterfaces()).any(new ClassPredicate(Collection.class, Iterable.class));
+                Arrays.stream(javaType.getInterfaces())
+                        .anyMatch(it -> matchesClass(it, Collection.class, Iterable.class));
     }
 
     private static List<?> convertToParameterizedJavaClass(final ParameterizedType javaType,
@@ -86,13 +83,8 @@ public class ListTypeCompatibility implements TypeCompatibility {
             return convertToJavaClass((Class) javaType.getRawType(), collection);
         }
         final Class<?> genericClass = (Class) javaType.getActualTypeArguments()[0];
-        return Iter.of(collection.getCollection())
-                .map(new Func<Object, LispValue>() {
-                    @Override
-                    public Object apply(final LispValue value) {
-                        return CallHelpers.convertValue(genericClass, value);
-                    }
-                }).toList();
+        return collection.stream()
+                .map(value -> CallHelpers.convertValue(genericClass, value)).collect(Collectors.toList());
     }
 
     @SuppressWarnings("unchecked")
@@ -101,7 +93,7 @@ public class ListTypeCompatibility implements TypeCompatibility {
         if (javaType.isArray()) {
             return (T) convertToArray(javaType, collection);
         } else {
-            return (T) Iter.of(collection.getCollection()).map(SIMPLE_VALUE_CONVERTER).toList();
+            return (T) collection.stream().map(SIMPLE_VALUE_CONVERTER).collect(Collectors.toList());
         }
     }
 
@@ -113,17 +105,8 @@ public class ListTypeCompatibility implements TypeCompatibility {
         return ArrayHelpers.convertToArray(elementType, values);
     }
 
-    private static final class ClassPredicate implements Predicate<Class<?>> {
-
-        private final Set<Class<?>> matches;
-
-        private ClassPredicate(final Class<?>... matches) {
-            this.matches = Collect.setOf(matches);
-        }
-
-        @Override
-        public boolean matches(final Class<?> aClass) {
-            return matches.contains(aClass);
-        }
+    private static boolean matchesClass(final Class<?> aClass, final Class<?>... matches) {
+        return ImmutableSet.copyOf(matches).contains(aClass);
     }
+
 }
